requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# AI/ML
google-generativeai==0.3.1

# Document Processing
PyPDF2==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# HTTP Client
httpx==0.25.1
aiofiles==23.2.1
requests==2.31.0
